Refactor performance test into own class and added documentation
